fixed a bug with User.legBettingTiles, deleted the abstract class Card and moved the id and color attributes into RaceBettingCard and LegBettingTile

+++ service/src/main/java/ch/uzh/ifi/seal/soprafs15/controller/GameServiceController.java
@@ -227,7 +227,7 @@ public class GameServiceController extends GenericService {
 	@RequestMapping(method = RequestMethod.GET, value = CONTEXT + "/{gameId}/players/{playerId}")
 	@ResponseStatus(HttpStatus.OK)
     @ResponseBody
-	public GamePlayerResponseBean getPlayer(@PathVariable Long gameId, @PathVariable Long playerId) {
+	public GamePlayerResponseBean getPlayer(@PathVariable Long gameId, @PathVariable Integer playerId) {
 		logger.debug("getPlayer: " + gameId);
 
         GamePlayerResponseBean result = gamePlayerService.getPlayer(gameId, playerId);
